refactor(core): split services module by responsibility
