+++ crates/loon-core/src/path/planner.rs
@@ -1,0 +1,672 @@
+use super::helpers::{
+    final_component, lookup_path, parse_absolute_path_for_core, parse_mutation_path,
+};
+use super::mutation::PutFileBehavior;
+use super::tombstone::reject_tombstoned_path_ancestor;
+use crate::basis::load_verified_namespace_basis;
+use crate::commit::SemanticCommitFingerprint;
+use crate::content::validate_durable_content_reference;
+use crate::error::CoreError;
+use crate::metadata::{MetadataState, ResolvedVisiblePath, VisiblePathError};
+use crate::publisher::{PathMutationIntent, PlannedPathMutation};
+use loon_api::{
+    payload_checksum_sha256,
+    v0::{
+        CommitOp as ApiCommitOp, CommitPrecondition as ApiCommitPrecondition,
+        CommitRequest as ApiCommitRequest, RenameMode,
+    },
+    AbsolutePath, ChangeSeq, CommitId, ContentRef, ContentStoreId, DisplayName, HeadState, InodeId,
+    InodeKind, NameKey, NamespaceId,
+};
+use loon_objectstore::ObjectStore;
+use serde::Serialize;
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize)]
+#[serde(tag = "type", rename_all = "snake_case")]
+enum PathFingerprintInput {
+    CreateDir {
+        namespace_id: NamespaceId,
+        absolute_path: String,
+    },
+    PutFile {
+        namespace_id: NamespaceId,
+        absolute_path: String,
+        behavior: PutFileBehavior,
+        content_ref: ContentRef,
+    },
+    DeletePath {
+        namespace_id: NamespaceId,
+        absolute_path: String,
+        recursive: bool,
+    },
+    MovePath {
+        namespace_id: NamespaceId,
+        from_path: String,
+        to_path: String,
+        mode: RenameMode,
+    },
+    CopyFilePath {
+        namespace_id: NamespaceId,
+        from_path: String,
+        to_path: String,
+    },
+}
+
+fn path_semantic_commit_fingerprint(
+    identity: &PathFingerprintInput,
+) -> Result<SemanticCommitFingerprint, CoreError> {
+    payload_checksum_sha256(identity)
+        .map(SemanticCommitFingerprint::new_unchecked)
+        .map_err(|err| CoreError::Store(err.to_string()))
+}
+
+pub(crate) fn semantic_commit_fingerprint_for_path_intent(
+    namespace_id: &NamespaceId,
+    intent: &PathMutationIntent,
+) -> Result<SemanticCommitFingerprint, CoreError> {
+    let identity = match intent {
+        PathMutationIntent::CreateDir { absolute_path, .. } => PathFingerprintInput::CreateDir {
+            namespace_id: namespace_id.clone(),
+            absolute_path: normalized_path_for_fingerprint(absolute_path)?,
+        },
+        PathMutationIntent::PutFile {
+            absolute_path,
+            behavior,
+            content_ref,
+            ..
+        } => PathFingerprintInput::PutFile {
+            namespace_id: namespace_id.clone(),
+            absolute_path: normalized_path_for_fingerprint(absolute_path)?,
+            behavior: *behavior,
+            content_ref: content_ref.clone(),
+        },
+        PathMutationIntent::DeletePath {
+            absolute_path,
+            recursive,
+            ..
+        } => PathFingerprintInput::DeletePath {
+            namespace_id: namespace_id.clone(),
+            absolute_path: normalized_path_for_fingerprint(absolute_path)?,
+            recursive: *recursive,
+        },
+        PathMutationIntent::MovePath {
+            from_path,
+            to_path,
+            mode,
+            ..
+        } => PathFingerprintInput::MovePath {
+            namespace_id: namespace_id.clone(),
+            from_path: normalized_path_for_fingerprint(from_path)?,
+            to_path: normalized_path_for_fingerprint(to_path)?,
+            mode: *mode,
+        },
+        PathMutationIntent::CopyFilePath {
+            from_path, to_path, ..
+        } => PathFingerprintInput::CopyFilePath {
+            namespace_id: namespace_id.clone(),
+            from_path: normalized_path_for_fingerprint(from_path)?,
+            to_path: normalized_path_for_fingerprint(to_path)?,
+        },
+    };
+    path_semantic_commit_fingerprint(&identity)
+}
+
+fn normalized_path_for_fingerprint(absolute_path: &str) -> Result<String, CoreError> {
+    Ok(parse_absolute_path_for_core(absolute_path)?
+        .as_str()
+        .to_owned())
+}
+
+pub(crate) fn plan_path_mutation<S: ObjectStore + ?Sized>(
+    store: &S,
+    namespace_id: &NamespaceId,
+    intent: &PathMutationIntent,
+) -> Result<PlannedPathMutation, CoreError> {
+    let basis = load_verified_namespace_basis(store, namespace_id)?;
+    plan_path_mutation_against_state(
+        store,
+        namespace_id,
+        intent,
+        &basis.head,
+        &basis.metadata_state,
+        &basis.content_store_id,
+    )
+}
+
+pub(crate) fn plan_path_mutation_against_state<S: ObjectStore + ?Sized>(
+    store: &S,
+    namespace_id: &NamespaceId,
+    intent: &PathMutationIntent,
+    head: &HeadState,
+    metadata_state: &MetadataState,
+    content_store_id: &ContentStoreId,
+) -> Result<PlannedPathMutation, CoreError> {
+    let commit_id = intent.commit_id().clone();
+    let semantic_fingerprint = semantic_commit_fingerprint_for_path_intent(namespace_id, intent)?;
+    let view = PathPlanningView {
+        head,
+        metadata_state,
+        content_store_id,
+    };
+    let commit_request = match intent {
+        PathMutationIntent::CreateDir { absolute_path, .. } => {
+            plan_create_dir(absolute_path, &commit_id, &view)?
+        }
+        PathMutationIntent::PutFile {
+            absolute_path,
+            content_ref,
+            behavior,
+            ..
+        } => plan_put_file_content_ref(
+            store,
+            absolute_path,
+            content_ref.clone(),
+            *behavior,
+            &commit_id,
+            &view,
+        )?,
+        PathMutationIntent::DeletePath {
+            absolute_path,
+            recursive,
+            ..
+        } => plan_delete_path(absolute_path, *recursive, &commit_id, &view)?,
+        PathMutationIntent::MovePath {
+            from_path,
+            to_path,
+            mode,
+            ..
+        } => plan_move_path(from_path, to_path, *mode, &commit_id, &view)?,
+        PathMutationIntent::CopyFilePath {
+            from_path, to_path, ..
+        } => plan_copy_file_path(store, from_path, to_path, &commit_id, &view)?,
+    };
+    Ok(PlannedPathMutation {
+        commit_id,
+        semantic_commit_fingerprint: semantic_fingerprint,
+        commit_request,
+    })
+}
+
+struct PathPlanningView<'a> {
+    head: &'a HeadState,
+    metadata_state: &'a MetadataState,
+    content_store_id: &'a ContentStoreId,
+}
+
+fn binding_is_precondition(
+    view: &PathPlanningView<'_>,
+    resolved: &ResolvedVisiblePath,
+) -> Result<ApiCommitPrecondition, CoreError> {
+    let parent_inode = resolved
+        .parent_inode_id
+        .ok_or(CoreError::RootMutationForbidden)?;
+    let binding = view
+        .metadata_state
+        .current_parent_binding_for_child(resolved.inode_id, view.head.seq)
+        .ok_or_else(|| CoreError::MissingPath(resolved.absolute_path.clone()))?;
+    if binding.parent_inode_id != parent_inode {
+        return Err(CoreError::MissingPath(resolved.absolute_path.clone()));
+    }
+    Ok(ApiCommitPrecondition::BindingIs {
+        parent_inode,
+        name_key: binding.name_key,
+        child_inode: binding.child_inode_id,
+        bind_seq: binding.bind_seq,
+        bind_delta_index: binding.bind_delta_index,
+    })
+}
+
+fn child_name_absent_precondition(
+    view: &PathPlanningView<'_>,
+    parent_inode: InodeId,
+    display_name: &str,
+) -> ApiCommitPrecondition {
+    let display_name =
+        DisplayName::parse(display_name).expect("path planner should provide valid display name");
+    let name_key = NameKey::for_display_name(view.head.name_policy, &display_name);
+    ApiCommitPrecondition::ChildNameAbsent {
+        parent_inode,
+        name_key: name_key.as_str().to_owned(),
+    }
+}
+
+fn plan_create_dir(
+    absolute_path: &str,
+    commit_id: &CommitId,
+    view: &PathPlanningView<'_>,
+) -> Result<ApiCommitRequest, CoreError> {
+    let absolute_path = parse_mutation_path(absolute_path)?;
+    reject_tombstoned_path_ancestor(
+        view.metadata_state,
+        &absolute_path,
+        view.head.name_policy,
+        view.head.seq,
+    )?;
+    if lookup_path(
+        view.metadata_state,
+        &absolute_path,
+        view.head.name_policy,
+        view.head.seq,
+    )
+    .is_ok()
+    {
+        return Err(CoreError::DestinationExists(
+            absolute_path.as_str().to_owned(),
+        ));
+    }
+    let parent_inode = resolve_parent_directory(
+        view.metadata_state,
+        &absolute_path,
+        view.head.name_policy,
+        view.head.seq,
+    )?;
+    let display_name = final_component(&absolute_path)?;
+    Ok(ApiCommitRequest {
+        commit_id: commit_id.to_owned(),
+        ops: vec![ApiCommitOp::CreateDir {
+            parent_inode,
+            display_name: display_name.clone(),
+        }],
+        preconditions: vec![
+            child_name_absent_precondition(view, parent_inode, &display_name),
+            ApiCommitPrecondition::AncestorsNotSubtreeDeleted {
+                inode_id: parent_inode,
+            },
+        ],
+        message: None,
+        annotations: None,
+    })
+}
+
+fn plan_put_file_content_ref<S: ObjectStore + ?Sized>(
+    store: &S,
+    absolute_path: &str,
+    content_ref: ContentRef,
+    behavior: PutFileBehavior,
+    commit_id: &CommitId,
+    view: &PathPlanningView<'_>,
+) -> Result<ApiCommitRequest, CoreError> {
+    let absolute_path = parse_mutation_path(absolute_path)?;
+    let _validated =
+        validate_durable_content_reference(store, view.content_store_id, &content_ref)?;
+    reject_tombstoned_path_ancestor(
+        view.metadata_state,
+        &absolute_path,
+        view.head.name_policy,
+        view.head.seq,
+    )?;
+    let target = lookup_path(
+        view.metadata_state,
+        &absolute_path,
+        view.head.name_policy,
+        view.head.seq,
+    );
+
+    let mut ops = Vec::new();
+    let mut working = view.metadata_state.clone();
+    let mut next_inode_id = view.head.next_inode_id;
+    let mut op_index = 0u32;
+    let final_parent_inode = ensure_parent_directories(
+        &absolute_path,
+        view.head.seq,
+        view.head.name_policy,
+        &mut working,
+        &mut ops,
+        &mut next_inode_id,
+        &mut op_index,
+    )?;
+    let final_name = final_component(&absolute_path)?;
+    let mut preconditions = Vec::new();
+
+    match target {
+        Ok(existing) => {
+            if behavior == PutFileBehavior::CreateOnly {
+                return Err(CoreError::DestinationExists(
+                    absolute_path.as_str().to_owned(),
+                ));
+            }
+            if existing.inode_kind != InodeKind::File {
+                return Err(CoreError::ExpectedFile {
+                    path: absolute_path.as_str().to_owned(),
+                    kind: existing.inode_kind,
+                });
+            }
+            let revision = view
+                .metadata_state
+                .latest_revision_head_at_seq(existing.inode_id, view.head.seq)
+                .ok_or_else(|| CoreError::MissingPath(absolute_path.as_str().to_owned()))?;
+            preconditions.push(binding_is_precondition(view, &existing)?);
+            ops.push(ApiCommitOp::ReplaceFile {
+                inode_id: existing.inode_id,
+                base_revision_no: revision.revision_no,
+                content_ref: content_ref.clone(),
+            });
+            preconditions.push(ApiCommitPrecondition::InodeRevisionIs {
+                inode_id: existing.inode_id,
+                revision_no: revision.revision_no,
+            });
+            preconditions.push(ApiCommitPrecondition::AncestorsNotSubtreeDeleted {
+                inode_id: existing.inode_id,
+            });
+        }
+        Err(VisiblePathError::PathNotFound { .. }) => {
+            ops.push(ApiCommitOp::CreateFile {
+                parent_inode: final_parent_inode,
+                display_name: final_name.clone(),
+                content_ref,
+            });
+            if view
+                .metadata_state
+                .visible_inode(final_parent_inode, view.head.seq)
+                .is_some()
+            {
+                preconditions.push(child_name_absent_precondition(
+                    view,
+                    final_parent_inode,
+                    &final_name,
+                ));
+                preconditions.push(ApiCommitPrecondition::AncestorsNotSubtreeDeleted {
+                    inode_id: final_parent_inode,
+                });
+            }
+        }
+        Err(other) => return Err(other.into()),
+    }
+
+    Ok(ApiCommitRequest {
+        commit_id: commit_id.to_owned(),
+        ops,
+        preconditions,
+        message: None,
+        annotations: None,
+    })
+}
+
+fn plan_delete_path(
+    absolute_path: &str,
+    recursive: bool,
+    commit_id: &CommitId,
+    view: &PathPlanningView<'_>,
+) -> Result<ApiCommitRequest, CoreError> {
+    let absolute_path = parse_mutation_path(absolute_path)?;
+    let resolved = view.metadata_state.resolve_visible_path(
+        &absolute_path,
+        view.head.name_policy,
+        view.head.seq,
+    )?;
+    let op = match resolved.inode_kind {
+        InodeKind::File => ApiCommitOp::DeleteFile {
+            inode_id: resolved.inode_id,
+        },
+        InodeKind::Dir if recursive => ApiCommitOp::DeleteSubtree {
+            root_inode: resolved.inode_id,
+        },
+        InodeKind::Dir => {
+            let children = view
+                .metadata_state
+                .visible_children(resolved.inode_id, view.head.seq);
+            if !children.is_empty() {
+                return Err(CoreError::DirectoryNotEmpty(
+                    absolute_path.as_str().to_owned(),
+                ));
+            }
+            ApiCommitOp::DeleteSubtree {
+                root_inode: resolved.inode_id,
+            }
+        }
+    };
+    let mut preconditions = vec![binding_is_precondition(view, &resolved)?];
+    if !recursive && resolved.inode_kind == InodeKind::Dir {
+        preconditions.push(ApiCommitPrecondition::DirectoryEmpty {
+            inode_id: resolved.inode_id,
+        });
+    }
+    Ok(ApiCommitRequest {
+        commit_id: commit_id.to_owned(),
+        ops: vec![op],
+        preconditions,
+        message: None,
+        annotations: None,
+    })
+}
+
+fn plan_move_path(
+    from_path: &str,
+    to_path: &str,
+    mode: RenameMode,
+    commit_id: &CommitId,
+    view: &PathPlanningView<'_>,
+) -> Result<ApiCommitRequest, CoreError> {
+    if mode != RenameMode::NoReplace {
+        return Err(CoreError::CommitValidation(
+            crate::commit::CommitValidationError::UnsupportedRenameMode { mode },
+        ));
+    }
+    let from_path = parse_mutation_path(from_path)?;
+    let to_path = parse_mutation_path(to_path)?;
+    reject_tombstoned_path_ancestor(
+        view.metadata_state,
+        &from_path,
+        view.head.name_policy,
+        view.head.seq,
+    )?;
+    reject_tombstoned_path_ancestor(
+        view.metadata_state,
+        &to_path,
+        view.head.name_policy,
+        view.head.seq,
+    )?;
+    let source = view.metadata_state.resolve_visible_path(
+        &from_path,
+        view.head.name_policy,
+        view.head.seq,
+    )?;
+    let target_parent = resolve_parent_directory(
+        view.metadata_state,
+        &to_path,
+        view.head.name_policy,
+        view.head.seq,
+    )?;
+    let target_name = final_component(&to_path)?;
+    if lookup_path(
+        view.metadata_state,
+        &to_path,
+        view.head.name_policy,
+        view.head.seq,
+    )
+    .is_ok()
+    {
+        return Err(CoreError::DestinationExists(to_path.as_str().to_owned()));
+    }
+    Ok(ApiCommitRequest {
+        commit_id: commit_id.to_owned(),
+        ops: vec![ApiCommitOp::Rename {
+            inode_id: source.inode_id,
+            new_parent_inode: target_parent,
+            new_display_name: target_name.clone(),
+            mode,
+        }],
+        preconditions: vec![
+            binding_is_precondition(view, &source)?,
+            child_name_absent_precondition(view, target_parent, &target_name),
+            ApiCommitPrecondition::AncestorsNotSubtreeDeleted {
+                inode_id: source.inode_id,
+            },
+            ApiCommitPrecondition::AncestorsNotSubtreeDeleted {
+                inode_id: target_parent,
+            },
+        ],
+        message: None,
+        annotations: None,
+    })
+}
+
+fn plan_copy_file_path<S: ObjectStore + ?Sized>(
+    store: &S,
+    from_path: &str,
+    to_path: &str,
+    commit_id: &CommitId,
+    view: &PathPlanningView<'_>,
+) -> Result<ApiCommitRequest, CoreError> {
+    let from_path = parse_mutation_path(from_path)?;
+    let to_path = parse_mutation_path(to_path)?;
+    reject_tombstoned_path_ancestor(
+        view.metadata_state,
+        &from_path,
+        view.head.name_policy,
+        view.head.seq,
+    )?;
+    reject_tombstoned_path_ancestor(
+        view.metadata_state,
+        &to_path,
+        view.head.name_policy,
+        view.head.seq,
+    )?;
+
+    let source = view.metadata_state.resolve_visible_path(
+        &from_path,
+        view.head.name_policy,
+        view.head.seq,
+    )?;
+    if source.inode_kind != InodeKind::File {
+        return Err(CoreError::ExpectedFile {
+            path: from_path.as_str().to_owned(),
+            kind: source.inode_kind,
+        });
+    }
+
+    if lookup_path(
+        view.metadata_state,
+        &to_path,
+        view.head.name_policy,
+        view.head.seq,
+    )
+    .is_ok()
+    {
+        return Err(CoreError::DestinationExists(to_path.as_str().to_owned()));
+    }
+
+    let revision = view
+        .metadata_state
+        .latest_revision_head_at_seq(source.inode_id, view.head.seq)
+        .ok_or_else(|| CoreError::MissingPath(from_path.as_str().to_owned()))?;
+    let _validated =
+        validate_durable_content_reference(store, view.content_store_id, &revision.content_ref)?;
+
+    let target_parent = resolve_parent_directory(
+        view.metadata_state,
+        &to_path,
+        view.head.name_policy,
+        view.head.seq,
+    )?;
+    let target_name = final_component(&to_path)?;
+    Ok(ApiCommitRequest {
+        commit_id: commit_id.to_owned(),
+        ops: vec![ApiCommitOp::CreateFile {
+            parent_inode: target_parent,
+            display_name: target_name.clone(),
+            content_ref: revision.content_ref,
+        }],
+        preconditions: vec![
+            binding_is_precondition(view, &source)?,
+            ApiCommitPrecondition::InodeRevisionIs {
+                inode_id: source.inode_id,
+                revision_no: revision.revision_no,
+            },
+            ApiCommitPrecondition::AncestorsNotSubtreeDeleted {
+                inode_id: source.inode_id,
+            },
+            child_name_absent_precondition(view, target_parent, &target_name),
+            ApiCommitPrecondition::AncestorsNotSubtreeDeleted {
+                inode_id: target_parent,
+            },
+        ],
+        message: None,
+        annotations: None,
+    })
+}
+
+fn ensure_parent_directories(
+    absolute_path: &AbsolutePath,
+    committed_seq: ChangeSeq,
+    name_policy: loon_api::NamePolicy,
+    working: &mut MetadataState,
+    ops: &mut Vec<ApiCommitOp>,
+    next_inode_id: &mut InodeId,
+    op_index: &mut u32,
+) -> Result<InodeId, CoreError> {
+    let components = absolute_path.components();
+    if components.len() <= 1 {
+        return Ok(InodeId(1));
+    }
+
+    let mut current_inode = InodeId(1);
+    for component in &components[..components.len() - 1] {
+        let display_name = component.to_display_name();
+        let name_key = NameKey::for_display_name(name_policy, &display_name);
+        if let Some(child) = working.visible_child(current_inode, name_key.as_str(), committed_seq)
+        {
+            let inode = working
+                .visible_inode(child.child_inode_id, committed_seq)
+                .ok_or_else(|| CoreError::MissingPath(component.as_str().to_owned()))?;
+            if inode.inode_kind != InodeKind::Dir {
+                return Err(CoreError::NonDirectoryPathComponent(
+                    component.as_str().to_owned(),
+                ));
+            }
+            current_inode = child.child_inode_id;
+            continue;
+        }
+
+        ops.push(ApiCommitOp::CreateDir {
+            parent_inode: current_inode,
+            display_name: display_name.as_str().to_owned(),
+        });
+        let allocated = *next_inode_id;
+        *next_inode_id = InodeId(next_inode_id.0.saturating_add(1));
+        let delta_index = op_index.saturating_mul(2);
+        let applied = working.apply_committed_wal_deltas(
+            committed_seq,
+            &[
+                loon_api::WalDelta::CreateInode {
+                    delta_index,
+                    inode_id: allocated,
+                    inode_kind: InodeKind::Dir,
+                },
+                loon_api::WalDelta::BindDirentry {
+                    delta_index: delta_index.saturating_add(1),
+                    parent_inode: current_inode,
+                    name_key: name_key.as_str().to_owned(),
+                    display_name: display_name.as_str().to_owned(),
+                    child_inode: allocated,
+                },
+            ],
+        )?;
+        *working = applied.metadata_state;
+        *op_index = op_index.saturating_add(1);
+        current_inode = allocated;
+    }
+    Ok(current_inode)
+}
+
+fn resolve_parent_directory(
+    metadata_state: &MetadataState,
+    absolute_path: &AbsolutePath,
+    name_policy: loon_api::NamePolicy,
+    seq: ChangeSeq,
+) -> Result<InodeId, CoreError> {
+    let Some(parent_path) = absolute_path.parent() else {
+        return Ok(InodeId(1));
+    };
+    if parent_path.is_root() {
+        return Ok(InodeId(1));
+    }
+    let resolved = metadata_state.resolve_visible_path(&parent_path, name_policy, seq)?;
+    if resolved.inode_kind != InodeKind::Dir {
+        return Err(CoreError::ExpectedDirectory {
+            path: parent_path.as_str().to_owned(),
+            kind: resolved.inode_kind,
+        });
+    }
+    Ok(resolved.inode_id)
+}
