+++ crates/loon-core/src/services/content.rs
@@ -1,0 +1,36 @@
+use crate::content::write_immutable_object;
+use crate::error::CoreError;
+use crate::namespace::catalog::load_namespace_content_store_id;
+use loon_api::{ContentRef, ContentStoreId, NamespaceId};
+use loon_objectstore::keys::content_blob;
+use loon_objectstore::ObjectStore;
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct StoredContent {
+    pub content_store_id: ContentStoreId,
+    pub object_key: String,
+    pub content_ref: ContentRef,
+    pub file_digest_sha256: String,
+    pub file_size_bytes: u64,
+}
+
+pub fn store_bytes_as_content<S: ObjectStore + ?Sized>(
+    store: &S,
+    namespace_id: &NamespaceId,
+    bytes: &[u8],
+) -> Result<StoredContent, CoreError> {
+    let content_store_id = load_namespace_content_store_id(store, namespace_id)?;
+    let content_ref = ContentRef::whole_file_v0(bytes);
+    let object_key = content_blob(content_store_id.as_str(), &content_ref.digest)
+        .map_err(|err| CoreError::Store(err.to_string()))?;
+    write_immutable_object(store, &object_key, bytes)?;
+
+    Ok(StoredContent {
+        content_store_id,
+        object_key,
+        file_digest_sha256: content_ref.digest.clone(),
+        file_size_bytes: content_ref.size_bytes,
+        content_ref,
+    })
+}
